Add material on interpreting level of significance and power
Added a new lecture that covers topics related to level of significance, and how repeated experiments have the risk of a type I error. Additional material includes discussion of how the power of a study will also relate to false negative results.

WIP: add an interactive component at the end of the lecture that would allow folks to control the sample size, effect size, level of significance, and power for developing numerical intuition.